// libs/features/contributor/src/lib/shared/models/health-record-item.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */
import { BilingualText } from '@gosi-ui/core';

export class HealthRecordItem {
  healthRecordDesc: BilingualText = new BilingualText();
  healthRecordId: number = undefined;
  remarkType: string = undefined;
}